Move `assignDefaultValues()` call into `createGlimmerPlugin()`
