+++ hbs-minifier-plugin.js
@@ -7,6 +7,8 @@ const trailingWhiteSpace = /[ \t\r\n]+$/;
 const WHITESPACE = /^[ \t\r\n]+$/;
 
 function createGlimmerPlugin(config) {
+  config = assignDefaultValues(config);
+
   // in this stack we track the nodes that cause us to skip the minification
   // e.g. `{{#no-minify}} ... {{/no-minify}}` blocks or `<pre></pre>` tags
   // depending on the configuration
@@ -219,6 +221,19 @@ function shouldSkipClass(node, config) {
   }
 
   return !canTrimWhiteSpaceBasedOnClassNames(classAttrNode.value, config.classes);
+}
+
+function assignDefaultValues(config) {
+  config = config || {};
+  let elements = config.elements || ['pre'];
+  let classes = config.classes || [];
+  let components = config.components || ['no-minify'];
+
+  return {
+    elements,
+    classes,
+    components
+  };
 }
 
 module.exports = {
