+++ index.js
@@ -10,7 +10,7 @@ module.exports = {
     let options = app.options || {};
     let config = options['ember-hbs-minifier'] || {};
 
-    let HbsMinifierPlugin = require('./hbs-minifier-plugin').createRegistryPlugin(assignDefaultValues(config.skip));
+    let HbsMinifierPlugin = require('./hbs-minifier-plugin').createRegistryPlugin(config.skip);
     registry.add('htmlbars-ast-plugin', {
       name: 'hbs-minifier-plugin',
       plugin: HbsMinifierPlugin,
@@ -28,16 +28,3 @@ module.exports = {
     this._setupPreprocessorRegistry(app);
   }
 };
-
-function assignDefaultValues(config) {
-  config = config || {};
-  let elements = config.elements || ['pre'];
-  let classes = config.classes || [];
-  let components = config.components || ['no-minify'];
-
-  return {
-    elements,
-    classes,
-    components
-  };
-}
